+++ linkerd/protocol/http/src/e2e/scala/io/buoyant/linkerd/protocol/ResponseClassificationEndToEndTest.scala
@@ -55,9 +55,9 @@ class ResponseClassificationEndToEndTest extends FunSuite {
     req.host = "foo"
     await(client(req))
 
-    assert(stats.counters(Seq("http", "dst", "id", s"$$/inet/127.1/${downstream.port}", "path", "svc/foo", "success")) == 1)
-    assert(stats.counters(Seq("http", "dst", "path", "svc/foo", "success")) == 1)
-    assert(stats.counters(Seq("http", "srv", "127.0.0.1/0", "success")) == 1)
+    assert(stats.counters(Seq("http", "client", s"$$/inet/127.1/${downstream.port}", "service", "svc/foo", "success")) == 1)
+    assert(stats.counters(Seq("http", "service", "svc/foo", "success")) == 1)
+    assert(stats.counters(Seq("http", "server", "127.0.0.1/0", "success")) == 1)
   }
 
   test("non-retryable failure") {
@@ -89,9 +89,9 @@ class ResponseClassificationEndToEndTest extends FunSuite {
     req.host = "foo"
     await(client(req))
 
-    assert(stats.counters(Seq("http", "dst", "id", s"$$/inet/127.1/${downstream.port}", "path", "svc/foo", "failures")) == 1)
-    assert(stats.counters(Seq("http", "dst", "path", "svc/foo", "failures")) == 1)
-    assert(stats.counters(Seq("http", "srv", "127.0.0.1/0", "failures")) == 1)
+    assert(stats.counters(Seq("http", "client", s"$$/inet/127.1/${downstream.port}", "service", "svc/foo", "failures")) == 1)
+    assert(stats.counters(Seq("http", "service", "svc/foo", "failures")) == 1)
+    assert(stats.counters(Seq("http", "server", "127.0.0.1/0", "failures")) == 1)
   }
 
   test("retryable failure") {
@@ -128,12 +128,12 @@ class ResponseClassificationEndToEndTest extends FunSuite {
     req.host = "foo"
     await(client(req))
 
-    assert(stats.counters(Seq("http", "srv", "127.0.0.1/0", "success")) == 1)
-    assert(stats.counters.get(Seq("http", "srv", "127.0.0.1/0", "failures")) == None)
-    assert(stats.counters(Seq("http", "dst", "path", "svc/foo", "success")) == 1)
-    assert(stats.counters.get(Seq("http", "dst", "path", "svc/foo", "failures")) == None)
-    assert(stats.counters(Seq("http", "dst", "id", s"$$/inet/127.1/${downstream.port}", "path", "svc/foo", "success")) == 1)
-    assert(stats.counters(Seq("http", "dst", "id", s"$$/inet/127.1/${downstream.port}", "path", "svc/foo", "failures")) == 1)
+    assert(stats.counters(Seq("http", "server", "127.0.0.1/0", "success")) == 1)
+    assert(stats.counters.get(Seq("http", "server", "127.0.0.1/0", "failures")) == None)
+    assert(stats.counters(Seq("http", "service", "svc/foo", "success")) == 1)
+    assert(stats.counters.get(Seq("http", "service", "svc/foo", "failures")) == None)
+    assert(stats.counters(Seq("http", "client", s"$$/inet/127.1/${downstream.port}", "service", "svc/foo", "success")) == 1)
+    assert(stats.counters(Seq("http", "client", s"$$/inet/127.1/${downstream.port}", "service", "svc/foo", "failures")) == 1)
   }
 
   test("per service classification") {
@@ -178,23 +178,23 @@ class ResponseClassificationEndToEndTest extends FunSuite {
     req.host = "a"
     await(client(req))
 
-    assert(stats.counters(Seq("http", "srv", "127.0.0.1/0", "success")) == 1)
-    assert(stats.counters.get(Seq("http", "srv", "127.0.0.1/0", "failures")) == None)
-    assert(stats.counters(Seq("http", "dst", "path", "svc/a", "success")) == 1)
-    assert(stats.counters.get(Seq("http", "dst", "path", "svc/a", "failures")) == None)
-    assert(stats.counters(Seq("http", "dst", "id", s"$$/inet/127.1/${downstream.port}", "path", "svc/a", "success")) == 1)
-    assert(stats.counters(Seq("http", "dst", "id", s"$$/inet/127.1/${downstream.port}", "path", "svc/a", "failures")) == 1)
+    assert(stats.counters(Seq("http", "server", "127.0.0.1/0", "success")) == 1)
+    assert(stats.counters.get(Seq("http", "server", "127.0.0.1/0", "failures")) == None)
+    assert(stats.counters(Seq("http", "service", "svc/a", "success")) == 1)
+    assert(stats.counters.get(Seq("http", "service", "svc/a", "failures")) == None)
+    assert(stats.counters(Seq("http", "client", s"$$/inet/127.1/${downstream.port}", "service", "svc/a", "success")) == 1)
+    assert(stats.counters(Seq("http", "client", s"$$/inet/127.1/${downstream.port}", "service", "svc/a", "failures")) == 1)
 
     // Request to "b" is non-retryable
     // fails and doesn't retry
     req.host = "b"
     await(client(req))
 
-    assert(stats.counters(Seq("http", "srv", "127.0.0.1/0", "success")) == 1)
-    assert(stats.counters(Seq("http", "srv", "127.0.0.1/0", "failures")) == 1)
-    assert(stats.counters.get(Seq("http", "dst", "path", "svc/b", "success")) == None)
-    assert(stats.counters(Seq("http", "dst", "path", "svc/b", "failures")) == 1)
-    assert(stats.counters.get(Seq("http", "dst", "id", s"$$/inet/127.1/${downstream.port}", "path", "svc/b", "success")) == None)
-    assert(stats.counters(Seq("http", "dst", "id", s"$$/inet/127.1/${downstream.port}", "path", "svc/b", "failures")) == 1)
+    assert(stats.counters(Seq("http", "server", "127.0.0.1/0", "success")) == 1)
+    assert(stats.counters(Seq("http", "server", "127.0.0.1/0", "failures")) == 1)
+    assert(stats.counters.get(Seq("http", "service", "svc/b", "success")) == None)
+    assert(stats.counters(Seq("http", "service", "svc/b", "failures")) == 1)
+    assert(stats.counters.get(Seq("http", "client", s"$$/inet/127.1/${downstream.port}", "service", "svc/b", "success")) == None)
+    assert(stats.counters(Seq("http", "client", s"$$/inet/127.1/${downstream.port}", "service", "svc/b", "failures")) == 1)
   }
 }