+++ linkerd/protocol/http/src/main/scala/io/buoyant/linkerd/protocol/http/ResponseClassifiers.scala
@@ -1,5 +1,6 @@
 package io.buoyant.linkerd.protocol.http
 
+import com.twitter.finagle.buoyant.linkerd.Headers
 import com.twitter.finagle.service.ResponseClassifier
 import com.twitter.finagle.service.RetryPolicy.{TimeoutAndWriteExceptionsOnly, ChannelClosedExceptionsOnly}
 import com.twitter.finagle.http.{Method, Request, Response, Status}
@@ -98,12 +99,23 @@ object ResponseClassifiers {
     HttpResponseClassifier.ServerErrorsAsFailures
 
   def NonRetryableChunked(classifier: ResponseClassifier): ResponseClassifier =
-    ResponseClassifier.named(s"NonRetryableChunked") {
+    ResponseClassifier.named(s"NonRetryableChunked[$classifier]") {
       case rr@ReqRep(req, _) if classifier.isDefinedAt(rr) =>
         (req, classifier(rr)) match {
           case (req: Request, ResponseClass.RetryableFailure) if req.isChunked =>
             ResponseClass.NonRetryableFailure
           case (_, rc) => rc
+        }
+    }
+
+  def HeaderRetryable(classifier: ResponseClassifier): ResponseClassifier =
+    ResponseClassifier.named(s"HeaderRetryable[$classifier]") {
+      case rr@ReqRep(req, Return(rsp: Response)) if classifier.isDefinedAt(rr) =>
+        val rc = classifier(rr)
+        if (rc == ResponseClass.NonRetryableFailure && Headers.Retryable.get(rsp.headerMap)) {
+          ResponseClass.RetryableFailure
+        } else {
+          rc
         }
     }
 }