+++ router/core/src/test/scala/io/buoyant/router/PerDstPathStatsFilterTest.scala
@@ -44,7 +44,7 @@ class PerDstPathStatsFilterTest extends FunSuite {
     assert(await(service("cat").liftToTry).isThrow)
     await(service("dog"))
 
-    val pfx = Seq("pfx", "path")
+    val pfx = Seq("pfx", "service")
     val catPfx = pfx :+ "req/cat"
     val dogPfx = pfx :+ "req/dog"
     assert(stats.counters == Map(
@@ -60,8 +60,8 @@ class PerDstPathStatsFilterTest extends FunSuite {
       (dogPfx :+ "pending")
     ))
     assert(stats.histogramDetails.keys == Set(
-      "pfx/path/req/cat/request_latency_ms",
-      "pfx/path/req/dog/request_latency_ms"
+      "pfx/service/req/cat/request_latency_ms",
+      "pfx/service/req/dog/request_latency_ms"
     ))
   }