+++ router/http/src/e2e/scala/io/buoyant/router/HttpEndToEndTest.scala
@@ -158,7 +158,7 @@ class HttpEndToEndTest extends FunSuite with Awaits {
 
     val stats = new InMemoryStatsReceiver
     def downstreamCounter(name: String) = {
-      val k = Seq("http", "dst", "id", s"$$/inet/127.1/${downstream.port}", name)
+      val k = Seq("http", "client", s"$$/inet/127.1/${downstream.port}", name)
       stats.counters.get(k)
     }
 
@@ -239,11 +239,11 @@ class HttpEndToEndTest extends FunSuite with Awaits {
 
     val stats = new InMemoryStatsReceiver
     def downstreamCounter(name: String) = {
-      val k = Seq("http", "dst", "id", s"$$/inet/127.1/${downstream.port}", name)
+      val k = Seq("http", "client", s"$$/inet/127.1/${downstream.port}", name)
       stats.counters.get(k)
     }
     def serverCounter(name: String) = {
-      val k = Seq("srv", name)
+      val k = Seq("server", name)
       stats.counters.get(k)
     }
 
@@ -273,7 +273,7 @@ class HttpEndToEndTest extends FunSuite with Awaits {
         .configured(param.Stats(stats))
         .factory()
       Http.server
-        .configured(param.Stats(stats.scope("srv")))
+        .configured(param.Stats(stats.scope("server")))
         .serve(new InetSocketAddress(0), factory)
     }